+++ bundle/deploy/metadata/compute.go
@@ -35,7 +35,7 @@ func (m *compute) Apply(_ context.Context, b *bundle.Bundle) diag.Diagnostics {
 	}
 
 	// Set job config paths in metadata
-	jobsMetadata := make(map[string]*metadata.Job)
+	jobsMetadata := make(map[string]*metadata.Resource)
 	for name, job := range b.Config.Resources.Jobs {
 		// Compute config file path the job is defined in, relative to the bundle
 		// root
@@ -45,12 +45,30 @@ func (m *compute) Apply(_ context.Context, b *bundle.Bundle) diag.Diagnostics {
 			return diag.Errorf("failed to compute relative path for job %s: %v", name, err)
 		}
 		// Metadata for the job
-		jobsMetadata[name] = &metadata.Job{
+		jobsMetadata[name] = &metadata.Resource{
 			ID:           job.ID,
 			RelativePath: filepath.ToSlash(relativePath),
 		}
 	}
 	b.Metadata.Config.Resources.Jobs = jobsMetadata
+
+	// Set pipeline config paths in metadata
+	pipelinesMetadata := make(map[string]*metadata.Resource)
+	for name, pipeline := range b.Config.Resources.Pipelines {
+		// Compute config file path the pipeline is defined in, relative to the bundle
+		// root
+		l := b.Config.GetLocation("resources.pipelines." + name)
+		relativePath, err := filepath.Rel(b.BundleRootPath, l.File)
+		if err != nil {
+			return diag.Errorf("failed to compute relative path for pipeline %s: %v", name, err)
+		}
+		// Metadata for the pipeline
+		pipelinesMetadata[name] = &metadata.Resource{
+			ID:           pipeline.ID,
+			RelativePath: filepath.ToSlash(relativePath),
+		}
+	}
+	b.Metadata.Config.Resources.Pipelines = pipelinesMetadata
 
 	// Set file upload destination of the bundle in metadata
 	b.Metadata.Config.Workspace.FilePath = b.Config.Workspace.FilePath